File hierarchy according to Node.js conventions and added Browserify support
* Move 'libwebrtc' to 'third_party' folder
* Move 'build.js' to 'bin' folder
* Move Javascript binding files to 'lib' folder
* Changing testing command to standard 'npm test'
* Browserify support and pre-generated file for testing (dist/wrtc.js)
* Clean-up of examples using browser support files
* Removed all not commited files when doing 'make clean'

After all of this, I did a full 'make clean;npm install' and automatically it
downloaded and installed everything successfully, it only keeps to automate the
download and install of the 'depot tools' and the system dependencies.

It also keeps to unify the generation script, currently there is entries at
package.json, Makefile, build.js and binding.gyp, ideally there should be only
the last one as propossed on issue #47.