Add _GLIBCXX_USE_CXX11_ABI=0 to binding.gyp
This is necessary in order to support building node-webrtc with versions of gcc
newer than the version used to build libwebrtc.a. For example, the current
libwebrtc.a is built with 4.8.4. There are some very strange crashes and
undefined symbol errors that result if you then try to build node-webrtc with
5.4.0.