Run bridge example for Chrome and Firefox separately
The bridge example can only really support on RTCPeerConnection pair at a time,
due to how offers, answers, and ICE candidates are signaled over the WebSocket.
In the long-term, we need to improve the bridge example to support multiple
RTCPeerConnections at a time. In the short-term, we'll just run the bridge
example twice.