[bugfix] Got rid of `run-script-os`.
It is not reliable to use in in `preinstall` phase, because deps
may or may not exist depending on other modules.
OTOH, creating the link in `install` is too late because then the
link to `node_modules/.bin` is not made.

So implementing the symlinking in JS. That's portable, right?